transaction entity test

+++ services/src/main/java/com/heavyduty/services/repository/TransactionsRepository.java
@@ -5,7 +5,7 @@ import com.heavyduty.services.entities.AccountsEntity;
 import com.heavyduty.services.entities.TransactionsEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TransactionsRepository extends CrudRepository<TransactionsEntity, Long> {
+public interface TransactionsRepository extends CrudRepository<TransactionsEntity, Integer> {
 
 
 }
